Kana-shift + space -> shift + space (for use in IME).
Standard Naginata style only defines spacebar as kana-shift when held.
The ZMK implementation "&ng SPACE" is tap for space, hold for kana-shift.
It adds "&ng ENTER" which is tap for enter, hold for kana-shift. This is
an alternative to the 2-key combo for enter.

This change lets you hold "&ng ENTER" and tap "&ng SPACE" to get LS(SPACE)
for selecting the previous kanji candidate in the IME.

It also lets you hold &ng_SPACE and tap &ng_ENTER to get LS(ENTER), which
might be useful for messager apps etc? Or, this could also send LS(SPACE)?